added the AteInferenceObject as utils to model_utils
